Make scripts work in inserted HTML
I had done this for general pages, but not for elsewhere. Turns out we might want it everywhere.
